chore: files changed src/openhuman/modules/voice.rs,src/openhuman/modules/voice_tests.rs
Auto-committed-on: dragonfly
